Introduction of Moon Monitors - refactoring moon/main.py
